Refactor canvas and backdrop
