+++ src/Graph/Backdrop/Backdrop.styled.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Backdrop = styled.rect`
+    height: ${({ dimensions }) => `${dimensions.boundedHeight}px`};
+    width: ${({ dimensions }) => `${dimensions.boundedWidth}px`};
+    fill: ${({ theme }) => theme.white};
+`;
