+++ src/Graph/Backdrop/index.js
@@ -1,0 +1,3 @@
+import Backdrop from './Backdrop';
+
+export default Backdrop;
